+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.1 - 2025-04-10]
+## [1.1.1 - 2025-05-21]
 ### Added
+- WebClient Factory
 
+### Changed
+- Refactoring OpalEngine
+- Refactoring Opal Waiting
 
 ## [1.1.0 - 2025-04-10]
 ### Added
@@ -50,6 +54,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Container Attributes Comparator (for FHIR)
 - Logger in FhirContainerAttributesComparator
 - Escape Characters in Fhir Path Merger
+- WebClientFactory class
+- WebClientFactory Attribute and Method in OpalServer
+- ExporterConst for WebClient request-url
+- Log OpalEngine
+- RequestBody Class
+- Lombok in Opal View objects
 
 ### Changed
 - Move FHIR Packages to dktk-exporter
@@ -65,9 +75,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Zip filename with SITE and TIMESTAMP
 - Refactoring Dockerfile
 - OPAL CSV (replacement of quotes with escape charachters)
+- Replaced Python commands in OpalEngine with WebClient request
+- waitUntilTaskIsFinished Method in OpalEngine
+- Replace RequestBody Class with ImportPathBody, CreateProjectBody, CsvDatasource, Csv DatasourceTable
+- Changed ViewFactory to JSONFactory
+- Changed JSONFactory to OpalClientBodyFactory
 
 ### Removed
-- Remove retrieve-query
+- Retrieve-query
+- Unnecessary comments and methods in Opal Engine
+- Constants used for Python commands
 
 ### Fixed
 - JSON and XML filter
@@ -79,6 +96,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Rest service: Get output formats
 - Extract parameters of HttpServletRequest
 - Opal Views
+- Warnings in OpalEngine
 - Add missing fields in update query
 
 ## [1.0.0 - 2023-07-20]